refactor(json-lint): move code mirror lint to custom packages
