+++ src/components/customizers/CssCustomizer.js
@@ -1,10 +1,9 @@
 import Codemirror from 'codemirror';
-import 'jsonlint';
+import 'code-mirror-lint';
 import PropTypes from 'prop-types';
 import React, { useRef, useEffect, useState } from 'react';
 
 import { style } from '../templates/FolesTemplate/Styles';
-import './lint';
 import { editorOptions } from './options';
 
 export const CssCustomizer = ({ setStyle }) => {
