+++ src/components/customizers/JsonCustomizer.js
@@ -1,10 +1,9 @@
 import CodeMirror from 'codemirror';
-import 'jsonlint';
+import 'code-mirror-lint';
 import PropTypes from 'prop-types';
 import React, { useRef, useEffect, useState } from 'react';
 
 import data from '../../example-json/john_smith.json';
-import './lint';
 import { editorOptions } from './options';
 
 export const JsonCustomizer = ({ setData }) => {
